Allow model save nodes to be used without onnx or ncnn installed (#872)
* Allow model save nodes to be used without onnx or ncnn installed

* better return

* fix import

* remove dividing line
