+++ backend/src/nodes/ncnn_nodes.py
@@ -19,6 +19,10 @@ from .utils.ncnn_auto_split import ncnn_auto_split_process
 from .utils.ncnn_model import NcnnModel
 from .utils.utils import get_h_w_c, convenient_upscale
 
+# NCNN Save Model node
+# pylint: disable=unused-import
+from .model_save_nodes import NcnnSaveNode
+
 
 @NodeFactory.register("chainner:ncnn:load_model")
 class NcnnLoadModelNode(NodeBase):
@@ -38,38 +42,6 @@ class NcnnLoadModelNode(NodeBase):
         model_name = os.path.splitext(os.path.basename(param_path))[0]
 
         return model, model_name
-
-
-@NodeFactory.register("chainner:ncnn:save_model")
-class NcnnSaveNode(NodeBase):
-    def __init__(self):
-        super().__init__()
-        self.description = "Save an NCNN model to specified directory."
-        self.inputs = [
-            NcnnModelInput(),
-            DirectoryInput(has_handle=True),
-            TextInput("Param/Bin Name"),
-        ]
-        self.outputs = []
-
-        self.category = NCNNCategory
-        self.name = "Save Model"
-        self.icon = "MdSave"
-        self.sub = "Input & Output"
-
-        self.side_effects = True
-
-    def run(self, net: NcnnModel, directory: str, name: str) -> bool:
-        full_bin = f"{name}.bin"
-        full_param = f"{name}.param"
-        full_bin_path = os.path.join(directory, full_bin)
-        full_param_path = os.path.join(directory, full_param)
-
-        logger.info(f"Writing NCNN model to paths: {full_bin_path} {full_param_path}")
-        net.write_bin(full_bin_path)
-        net.write_param(full_param_path)
-
-        return True
 
 
 @NodeFactory.register("chainner:ncnn:upscale_image")
