+++ src/common/common-types.ts
@@ -165,4 +165,5 @@ export interface Category {
     icon: string;
     color: string;
     installHint?: string;
+    excludedFromCheck: string[];
 }
